requirements.txt
alembic==1.5.2
argon2-cffi==20.1.0
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
bcrypt==3.2.0
bleach==3.3.0
blinker==1.4
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
dnspython==2.1.0
dominate==2.6.0
email-validator==1.1.2
entrypoints==0.3
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Bootstrap==3.3.7.1
Flask-Dance==3.3.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.6.0
Flask-SQLAlchemy==2.4.4
Flask-talisman==0.7.0
Flask-WTF==0.14.3
future==0.18.2
gevent==21.1.2
glob2==0.7
google-api-core==1.26.0
google-api-python-client==1.12.8
google-auth==1.27.0
google-auth-httplib2==0.0.4
google-cloud-core==1.6.0
google-cloud-storage==1.36.0
google-crc32c==1.1.2
google-resumable-media==1.2.0
googleapis-common-protos==1.52.0
greenlet==1.0.0
gunicorn==20.0.4
httplib2==0.19.0
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.3
ipykernel==5.5.0
ipython==7.20.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.11
jupyter-console==6.2.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
lxml==4.6.2
Mako==1.1.4
MarkupSafe==1.1.1
matplotlib==3.3.3
mistune==0.8.4
nbclient==0.5.2
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
networkx==2.5
notebook==6.2.0
numpy==1.19.5
opencv-python==4.5.1.48
packaging==20.9
pandocfilters==1.4.3
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.0
proglog==0.1.9
prometheus-client==0.9.0
prompt-toolkit==3.0.16
protobuf==3.15.1
psycopg2==2.8.6
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.8.0
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2021.1
PyWavelets==1.1.1
pyzmq==22.0.3
qtconsole==5.0.2
QtPy==1.9.0
requests==2.25.1
rsa==4.7.1
scikit-image==0.17.2
scipy==1.5.4
scramp==1.2.0
Send2Trash==1.5.0
six==1.15.0
SQLAlchemy==1.3.22
sqlalchemy_utils==0.36.8
stripe==2.55.1
terminado==0.9.2
testpath==0.4.4
tifffile==2020.9.3
tornado==6.1
tqdm==4.56.0
traitlets==5.0.5
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.2
validators==0.18.2
visitor==0.1.3
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
WTForms==2.3.3
zope.event==4.5.0
zope.interface==5.2.0
google-cloud-secret-manager==2.0.0
flask-talisman==0.7.0
cryptography==3.3.1
WooCommerce==2.1.1
